[libc++][CI] Reorder jobs.
In the second leg of the CI the steps take about:
- C++2b              10m
- C++11              8m
- C++03              6m
- Modular build      10m
- GCC 12 / C++latest 20m
So the slowest job is scheduled last. The CI will wait to start the
third leg until that job is done. The current order increases the
latency of the current job, instead start the slow jobs earlier.

Reviewed By: #libc, ldionne

Differential Revision: https://reviews.llvm.org/D136276

NOKEYCHECK=True
GitOrigin-RevId: 37fc37dd5d4970d7aab70c7db3f9bee9bb042e2d